requirements.txt
absl-py==2.1.0
addict==2.4.0
aitviewer==1.13.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
argcomplete==3.5.3
attrs==25.1.0
av==14.0.1
beautifulsoup4==4.13.3
blinker==1.9.0
blis==1.2.0
blobfile==3.0.0
braceexpand==0.1.7
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.1
chumpy@git+https://github.com/mattloper/chumpy
click==8.1.8
clip@git+https://github.com/openai/CLIP.git
cloudpathlib==0.20.0
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
cython==3.0.11
defusedxml==0.7.1
dill==0.3.9
easydict==1.13
einops==0.8.0
evo==1.30.5
fastjsonschema==2.21.1
filelock==3.16.1
flask==3.1.0
fonttools==4.55.3
freetype-py==2.5.1
fsspec==2024.12.0
ftfy==6.3.1
gdown==5.2.0
glcontext==3.0.0
grpcio==1.70.0
h5py==3.13.0
html4vision==0.4.3
huggingface-hub==0.27.1
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
imgui==2.0.0
iopath==0.1.10
itsdangerous==2.2.0
jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langcodes==3.5.0
language-data==1.3.0
lapx==0.5.11.post1
lazy-loader==0.4
lightning-utilities==0.11.9
loguru==0.7.3
lxml==5.3.0
lz4==4.3.3
marisa-trie==1.2.1
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mmcv==1.3.9
mmengine==0.10.4
moderngl==5.12.0
moderngl-window==2.4.6
multipledispatch==1.0.0
murmurhash==1.0.12
narwhals==1.24.1
natsort==8.4.0
nbformat==5.10.4
networkx==3.4.2
numexpr==2.10.2
numpy==1.26.4
omegaconf==2.3.0
opencv-contrib-python-headless==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
plyfile==1.1
portalocker==3.1.1
preshed==3.0.9
pretty-errors==1.2.25
protobuf==5.29.3
psutil==6.1.1
pulp==2.9.0
py-cpuinfo==9.0.0
pycocotools==2.0.8
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic-core==2.27.2
pyglet==2.1.0
pyopengl==3.1.0
pyparsing==3.2.1
pyqt5==5.15.11
pyqt5-qt5==5.15.16
pyqt5-sip==12.16.1
pyrender==0.1.45
pyrootutils==1.0.4
pyrr==0.10.3
pysocks==1.7.1
python-dotenv==1.0.1
pytorch-lightning==2.2.4
pytorch3d@git+https://github.com/facebookresearch/pytorch3d.git
pytz==2024.2
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
roma==1.5.1
rosbags==0.10.7
rpds-py==0.22.3
ruamel-yaml==0.18.10
ruamel-yaml-clib==0.2.12
safetensors==0.5.2
scikit-image==0.25.0
scikit-video==1.1.11
scipy==1.15.1
seaborn==0.13.2
shellingham==1.5.4
smart-open==7.1.0
smplx==0.1.28
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
submitit==1.5.2
supervision==0.25.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
termcolor==2.5.0
thinc==8.3.4
tifffile==2025.1.10
timm==1.0.13
tomli==2.2.1
torch-scatter==2.1.2
torchmetrics==1.4.0
tqdm==4.67.1
trimesh==3.23.5
typer==0.15.1
tzdata==2024.2
ultralytics==8.3.62
ultralytics-thop==2.0.13
urllib3==2.3.0
usd-core==24.11
wasabi==1.1.3
weasel==0.4.1
webdataset==0.2.100
websockets==14.1
werkzeug==3.1.3
wrapt==1.17.2
xtcocotools==1.14.3
yacs==0.1.8
yapf==0.43.0
zstandard==0.23.0